+++ packages/swapper/src/swappers/zrx/ZrxSwapper.ts
@@ -5,6 +5,7 @@ import { KnownChainIds } from '@shapeshiftoss/types'
 import {
   ApprovalNeededInput,
   ApprovalNeededOutput,
+  ApproveAmountInput,
   ApproveInfiniteInput,
   BuildTradeInput,
   BuyAssetBySellIdInput,
@@ -16,14 +17,14 @@ import {
   SwapperType,
   TradeQuote,
   TradeResult,
-  TradeTxs
+  TradeTxs,
 } from '../../api'
 import { getZrxTradeQuote } from './getZrxTradeQuote/getZrxTradeQuote'
 import { ZrxExecuteTradeInput, ZrxSwapperDeps, ZrxTrade } from './types'
 import { UNSUPPORTED_ASSETS } from './utils/blacklist'
 import { getUsdRate } from './utils/helpers/helpers'
 import { zrxApprovalNeeded } from './zrxApprovalNeeded/zrxApprovalNeeded'
-import { zrxApproveInfinite } from './zrxApproveInfinite/zrxApproveInfinite'
+import { zrxApproveAmount, zrxApproveInfinite } from './zrxApprove/zrxApprove'
 import { zrxBuildTrade } from './zrxBuildTrade/zrxBuildTrade'
 import { zrxExecuteTrade } from './zrxExecuteTrade/zrxExecuteTrade'
 
@@ -45,7 +46,7 @@ export class ZrxSwapper<T extends EvmSupportedChainIds> implements Swapper<T> {
         return SwapperType.ZrxAvalanche
       default:
         throw new SwapError('[getType]', {
-          code: SwapErrorTypes.UNSUPPORTED_CHAIN
+          code: SwapErrorTypes.UNSUPPORTED_CHAIN,
         })
     }
   }
@@ -74,13 +75,17 @@ export class ZrxSwapper<T extends EvmSupportedChainIds> implements Swapper<T> {
     return zrxApproveInfinite<T>(this.deps, args)
   }
 
+  async approveAmount(args: ApproveAmountInput<T>): Promise<string> {
+    return zrxApproveAmount<T>(this.deps, args)
+  }
+
   filterBuyAssetsBySellAssetId(args: BuyAssetBySellIdInput): AssetId[] {
     const { assetIds = [], sellAssetId } = args
     return assetIds.filter(
       (id) =>
         id.startsWith(this.chainId) &&
         sellAssetId?.startsWith(this.chainId) &&
-        !UNSUPPORTED_ASSETS.includes(id)
+        !UNSUPPORTED_ASSETS.includes(id),
     )
   }
 
@@ -91,7 +96,7 @@ export class ZrxSwapper<T extends EvmSupportedChainIds> implements Swapper<T> {
   async getTradeTxs(tradeResult: TradeResult): Promise<TradeTxs> {
     return {
       sellTxid: tradeResult.tradeId,
-      buyTxid: tradeResult.tradeId
+      buyTxid: tradeResult.tradeId,
     }
   }
 }