+++ packages/unchained-client/src/evm/ethereum/parser/uniV2.ts
@@ -2,7 +2,7 @@ import { ChainId, fromChainId, toAssetId } from '@shapeshiftoss/caip'
 import { ethers } from 'ethers'
 
 import { Tx } from '../../../generated/ethereum'
-import { TransferType } from '../../../types'
+import { BaseTxMetadata, TransferType } from '../../../types'
 import { getSigHash, SubParser, txInteractsWithContract, TxSpecific } from '../../parser'
 import ERC20_ABI from '../../parser/abi/erc20'
 import UNIV2_ABI from './abi/uniV2'
@@ -11,8 +11,12 @@ import {
   UNI_V2_FOX_STAKING_REWARDS_V3,
   UNI_V2_ROUTER_CONTRACT,
   WETH_CONTRACT_MAINNET,
-  WETH_CONTRACT_ROPSTEN
+  WETH_CONTRACT_ROPSTEN,
 } from './constants'
+
+export interface TxMetadata extends BaseTxMetadata {
+  parser: 'uniV2'
+}
 
 export interface ParserArgs {
   chainId: ChainId
@@ -29,12 +33,12 @@ export class Parser implements SubParser<Tx> {
 
   readonly supportedFunctions = {
     addLiquidityEthSigHash: this.abiInterface.getSighash('addLiquidityETH'),
-    removeLiquidityEthSigHash: this.abiInterface.getSighash('removeLiquidityETH')
+    removeLiquidityEthSigHash: this.abiInterface.getSighash('removeLiquidityETH'),
   }
 
   readonly supportedStakingRewardsFunctions = {
     stakeSigHash: this.stakingRewardsInterface.getSighash('stake'),
-    exitSigHash: this.stakingRewardsInterface.getSighash('exit')
+    exitSigHash: this.stakingRewardsInterface.getSighash('exit'),
   }
 
   constructor(args: ParserArgs) {
@@ -81,7 +85,7 @@ export class Parser implements SubParser<Tx> {
           const assetId = toAssetId({
             ...fromChainId(this.chainId),
             assetNamespace: 'erc20',
-            assetReference: tokenAddress
+            assetReference: tokenAddress,
           })
 
           return [
@@ -92,8 +96,8 @@ export class Parser implements SubParser<Tx> {
               assetId,
               totalValue: value,
               components: [{ value }],
-              token: { contract: tokenAddress, decimals, name, symbol }
-            }
+              token: { contract: tokenAddress, decimals, name, symbol },
+            },
           ]
         }
         case this.supportedFunctions.removeLiquidityEthSigHash: {
@@ -106,7 +110,7 @@ export class Parser implements SubParser<Tx> {
           const assetId = toAssetId({
             ...fromChainId(this.chainId),
             assetNamespace: 'erc20',
-            assetReference: lpTokenAddress
+            assetReference: lpTokenAddress,
           })
 
           return [
@@ -117,8 +121,8 @@ export class Parser implements SubParser<Tx> {
               assetId,
               totalValue: value,
               components: [{ value }],
-              token: { contract: lpTokenAddress, decimals, name, symbol }
-            }
+              token: { contract: lpTokenAddress, decimals, name, symbol },
+            },
           ]
         }
         default:
@@ -133,8 +137,8 @@ export class Parser implements SubParser<Tx> {
       transfers,
       data: {
         parser: 'uniV2',
-        method: decoded.name
-      }
+        method: decoded.name,
+      },
     }
   }
 
@@ -154,8 +158,8 @@ export class Parser implements SubParser<Tx> {
     return {
       data: {
         parser: 'uniV2',
-        method: decoded.name
-      }
+        method: decoded.name,
+      },
     }
   }