+++ packages/swapper/CHANGELOG.md
@@ -1,3 +1,28 @@
+# [@shapeshiftoss/swapper-v9.16.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.16.1...@shapeshiftoss/swapper-v9.16.2) (2022-08-16)
+
+
+### Bug Fixes
+
+* off-by-one allowance check ([#979](https://github.com/shapeshift/lib/issues/979)) ([50dd012](https://github.com/shapeshift/lib/commit/50dd012b19da4b47d9b48750c708717982dde678))
+
+# [@shapeshiftoss/swapper-v9.16.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.16.0...@shapeshiftoss/swapper-v9.16.1) (2022-08-15)
+
+# [@shapeshiftoss/swapper-v9.16.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.15.1...@shapeshiftoss/swapper-v9.16.0) (2022-08-15)
+
+
+### Features
+
+* implement CowSwap approveAmount method ([#976](https://github.com/shapeshift/lib/issues/976)) ([e777dba](https://github.com/shapeshift/lib/commit/e777dbab4c1caebf62e6f0cca4e8547c356fc325))
+
+# [@shapeshiftoss/swapper-v9.15.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.15.0...@shapeshiftoss/swapper-v9.15.1) (2022-08-15)
+
+# [@shapeshiftoss/swapper-v9.15.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.14.0...@shapeshiftoss/swapper-v9.15.0) (2022-08-14)
+
+
+### Features
+
+* **swapper:** add approveAmount swapper API method ([#962](https://github.com/shapeshift/lib/issues/962)) ([8bf11b1](https://github.com/shapeshift/lib/commit/8bf11b15939d437b4fdd5668e75407d5ede931c8))
+
 # [@shapeshiftoss/swapper-v9.14.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/swapper-v9.13.0...@shapeshiftoss/swapper-v9.14.0) (2022-08-12)